requirements.txt
requests>=2.32.0
PyYAML>=6.0.2
python-dotenv>=1.0.1
typer>=0.12.3
pytest>=7.4.0
moviepy>=2.2.1
numpy>=2.3.5